+++ lib.esm/mnemonics/mnemonic.js
@@ -29,6 +29,9 @@ export class Mnemonic {
     getMnemonic() {
         return this.mnemonic.join(' ');
     }
+    getMnemonicType() {
+        throw new Error('Not implemented');
+    }
     getWords() {
         return this.words;
     }